+++ test/unit/Instana/SDK/Internal/IdTest.hs
@@ -1,9 +1,10 @@
+{-# LANGUAGE OverloadedStrings   #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 module Instana.SDK.Internal.IdTest (allTests) where
 
 
 import qualified Data.Aeson                 as Aeson
-import qualified Data.ByteString.Lazy.Char8 as BS
+import qualified Data.ByteString.Lazy.Char8 as LBSC8
 import           Test.HUnit
 
 import           Instana.SDK.Internal.Id    (Id)
@@ -13,7 +14,13 @@ import qualified Instana.SDK.Internal.Id    as Id
 allTests :: Test
 allTests =
   TestList
-    [ TestLabel "shouldEncodeIntComponents" shouldEncodeIntComponents
+    [ TestLabel
+        "shouldConvertIntComponentsToString"
+        shouldConvertIntComponentsToString
+    , TestLabel "shouldEncodeIntComponents" shouldEncodeIntComponents
+    , TestLabel
+        "shouldConvertIntComponentsToString2"
+        shouldConvertIntComponentsToString2
     , TestLabel "shouldEncodeIntComponents2" shouldEncodeIntComponents2
     , TestLabel "shouldEncodeStringId" shouldEncodeStringId
     , TestLabel "shouldDecodeStringId " shouldDecodeStringId
@@ -21,24 +28,43 @@ allTests =
     ]
 
 
+shouldConvertIntComponentsToString :: Test
+shouldConvertIntComponentsToString =
+  let
+    asString = Id.toString $ Id.createFromIntsForTest [12345, 67890]
+  in
+    TestCase $
+      assertEqual "as string" "0000303900010932" asString
+
+
 shouldEncodeIntComponents :: Test
 shouldEncodeIntComponents =
   let
     idFromInts = Id.createFromIntsForTest [12345, 67890]
-    encoded = BS.unpack . Aeson.encode $ idFromInts
+    encoded = LBSC8.unpack . Aeson.encode $ idFromInts
   in
     TestCase $
-      assertEqual "encoded" "\"0000303900010932\"" encoded
+      assertEqual "as string" "\"0000303900010932\"" encoded
+
+
+shouldConvertIntComponentsToString2 :: Test
+shouldConvertIntComponentsToString2 =
+  let
+    -- minimum for Int type as per
+    -- http://hackage.haskell.org/package/base-4.11.1.0/docs/Data-Int.html
+    -- is from -2^29 to 29-1 -> -536870912 to 536870911
+    asString =
+      Id.toString $ Id.createFromIntsForTest [-536870912, 536870911]
+  in
+    TestCase $
+      assertEqual "as string" "200000001fffffff" asString
 
 
 shouldEncodeIntComponents2 :: Test
 shouldEncodeIntComponents2 =
   let
-    -- minimum for Int type as per
-    -- http://hackage.haskell.org/package/base-4.11.1.0/docs/Data-Int.html
-    -- is from -2^29 to 29-1 -> -536870912 to 536870911
     idFromInts = Id.createFromIntsForTest [-536870912, 536870911]
-    encoded = BS.unpack . Aeson.encode $ idFromInts
+    encoded = LBSC8.unpack . Aeson.encode $ idFromInts
   in
     TestCase $
       assertEqual "encoded" "\"200000001fffffff\"" encoded
@@ -47,8 +73,8 @@ shouldEncodeIntComponents2 =
 shouldEncodeStringId :: Test
 shouldEncodeStringId =
   let
-    idFromString = Id.fromString "some string"
-    encoded = BS.unpack . Aeson.encode $ idFromString
+    idFromString = ("some string" :: Id)
+    encoded = LBSC8.unpack . Aeson.encode $ idFromString
   in
     TestCase $
       assertEqual "encoded" "\"some string\"" encoded
@@ -57,8 +83,8 @@ shouldEncodeStringId =
 shouldDecodeStringId :: Test
 shouldDecodeStringId =
   let
-    expected = Id.fromString "some string"
-    maybeId = Aeson.decode (BS.pack  "\"some string\"") :: Maybe Id
+    expected = ("some string" :: Id)
+    maybeId = Aeson.decode (LBSC8.pack  "\"some string\"") :: Maybe Id
     Just decoded = maybeId
   in
     TestCase $ assertEqual "decoded" expected decoded
@@ -70,7 +96,7 @@ shouldGenerate =
     do
       generated <- Id.generate
       let
-        encoded = BS.unpack . Aeson.encode $ generated
+        encoded = LBSC8.unpack . Aeson.encode $ generated
       -- every encoded ID must be 16 chars wide + 2 chars for leading and
       -- trailing "
       assertEqual "generated" 18 (length encoded)